refactor a bit the way images are concreting from data and type; improve naming; docs;
